Move zip extraction helper into utils
