+++ .changeset/qti-zip-utils.md
@@ -1,0 +1,5 @@
+---
+'@prairielearn/utils': minor
+---
+
+Add a Node-only `@prairielearn/utils/zip` subpath for zip extraction with entry-count, expanded-size, and symlink validation.
