+++ packages/utils/package.json
@@ -11,16 +11,27 @@
     "node": ">=24.0.0"
   },
   "main": "./dist/index.js",
+  "exports": {
+    ".": "./dist/index.js",
+    "./zip": "./dist/zip/index.js"
+  },
   "scripts": {
     "build": "tsgo",
     "dev": "tsgo --watch --preserveWatchOutput",
     "test": "vitest run --coverage"
   },
+  "dependencies": {
+    "extract-zip": "^2.0.1",
+    "yauzl": "^2.10.0"
+  },
   "devDependencies": {
     "@prairielearn/tsconfig": "workspace:^",
+    "@types/archiver": "^7.0.0",
     "@types/node": "^24.12.4",
+    "@types/yauzl": "^2.10.3",
     "@typescript/native-preview": "beta",
     "@vitest/coverage-v8": "^4.1.7",
+    "archiver": "^7.0.1",
     "vitest": "^4.1.7"
   }
 }
