--- packages/zip/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@prairielearn/zip",
-  "version": "1.0.0",
-  "type": "module",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/PrairieLearn/PrairieLearn.git",
-    "directory": "packages/zip"
-  },
-  "private": true,
-  "engines": {
-    "node": ">=24.0.0"
-  },
-  "main": "./dist/index.js",
-  "scripts": {
-    "build": "tsgo",
-    "dev": "tsgo --watch --preserveWatchOutput",
-    "test": "vitest run --coverage"
-  },
-  "dependencies": {
-    "extract-zip": "^2.0.1",
-    "yauzl": "^2.10.0"
-  },
-  "devDependencies": {
-    "@prairielearn/tsconfig": "workspace:^",
-    "@types/archiver": "^7.0.0",
-    "@types/node": "^24.12.4",
-    "@types/yauzl": "^2.10.3",
-    "@typescript/native-preview": "beta",
-    "@vitest/coverage-v8": "^4.1.7",
-    "archiver": "^7.0.1",
-    "vitest": "^4.1.7"
-  },
-  "sideEffects": false
-}
