--- packages/zip/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "@prairielearn/tsconfig/tsconfig.package-node.json"
-}
